Use build tags to integrate with Celo
Importing both `go-ethereum` and `celo-blockchain`
modules in the same time turned to be impossible
due to several C linker conflicts. Under the hood,
both modules use some C code which declares
identical symbols. Because of that Go compilation
fails with a
`ld: 34 duplicate symbols for architecture x86_64`
error. To solve that problem, Go build tags has
been leveraged. This way, only one of the modules
is imported during compile time.
