+++ cmd/ethereum_signing_test.go
@@ -1,3 +1,5 @@
+//+build !celo
+
 package cmd
 
 import (
